tested on 30000 frames

+++ model/interpolation.py
@@ -4,11 +4,12 @@ import torch
 
 class QuinticPolynomial:
   def __init__(self, initial_time, initial_pos, initial_vel, initial_acc, final_time, final_pos, final_vel, final_acc, device=torch.device("cpu")):
-    self.device=device
-    self.position_coeff_ = torch.zeros((6, 1), dtype=torch.float).to(device)
-    self.velocity_coeff_ = torch.zeros((6, 1), dtype=torch.float).to(device)
-    self.acceleration_coeff_ = torch.zeros((6, 1), dtype=torch.float).to(device)
-    self.time_varibles = torch.zeros((1, 6), dtype=torch.float).to(device)
+    self.device = device
+    self.position_coeff_ = torch.zeros((6, 1), dtype=torch.float64).to(device)
+    self.velocity_coeff_ = torch.zeros((6, 1), dtype=torch.float64).to(device)
+    self.acceleration_coeff_ = torch.zeros((6, 1), dtype=torch.float64
+                                           ).to(device)
+    self.time_varibles = torch.zeros((1, 6), dtype=torch.float64).to(device)
 
     if final_time > initial_time:
       self.initial_time = initial_time
@@ -33,9 +34,9 @@ class QuinticPolynomial:
          [self.final_time ** 5, self.final_time ** 4, self.final_time ** 3, self.final_time ** 2, self.final_time, 1.0],
          [5.0 * self.final_time ** 4, 4.0 * self.final_time ** 3, 3.0 * self.final_time ** 2, 2.0 * self.final_time, 1.0, 0.0],
          [20.0 * self.final_time ** 3, 12.0 * self.final_time ** 2, 6.0 * self.final_time, 2.0, 0.0, 0.0]
-         ], dtype=torch.float).to(device)
+         ], dtype=torch.float64).to(device)
       self.conditions_mat = torch.tensor([self.initial_pos, self.initial_vel, self.initial_acc, self.final_pos, self.final_vel, self.final_acc],
-                                         dtype=torch.float).to(device)
+                                         dtype=torch.float64).to(device)
       self.conditions_mat = self.conditions_mat.reshape((6, 1))
       self.position_coeff_ = torch.linalg.pinv(self.time_mat) @ self.conditions_mat
       self.velocity_coeff_ = torch.tensor([[0.0],
@@ -68,7 +69,7 @@ class QuinticPolynomial:
       return self.current_pos
     else:
       self.current_time = t
-      self.time_varibles = torch.tensor([t ** 5, t ** 4, t ** 3, t ** 2, t, 1.0], dtype=torch.float).to(self.device)
+      self.time_varibles = torch.tensor([t ** 5, t ** 4, t ** 3, t ** 2, t, 1.0], dtype=torch.float64).to(self.device)
       self.current_pos = (self.time_varibles @ self.position_coeff_)[0]
       self.current_vel = (self.time_varibles @ self.velocity_coeff_)[0]
       self.current_acc = (self.time_varibles @ self.acceleration_coeff_)[0]
